pyproject.toml
[tool.poetry]
name = "openstates-scrapers"
version = "2020.4"
description = ""
authors = ["James Turk <james@openstates.org>"]

[tool.poetry.dependencies]
python = "^3.7"
xlrd = "<2"
lxml = "^4.4"
pytz = "^2019.3"
feedparser = "^5.2"
python-dateutil = "^2.8"
spatula = "^0.0.2"
chardet = "^3.0"
suds-py3 = "^1.3"
requests = {version = "^2.22", extras = ["security"]}
openstates-metadata = "*"
openstates = "^4.0"
mysqlclient = {version = "^1.4.6", optional=true}
SQLAlchemy = {version = "^1.3", optional=true}

[tool.poetry.extras]
california = ["mysqlclient", "SQLAlchemy"]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
